refactor: splitted auth; used tokens lib
